+++ src/main/java/chess/domain/board/BoardFactory.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
