+++ src/test/java/chess/ChessGameTest.java
@@ -9,8 +9,8 @@ import static chess.domain.piece.PieceType.QUEEN;
 import static chess.domain.piece.PieceType.ROOK;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.Board;
-import chess.domain.BoardFactory;
+import chess.domain.board.Board;
+import chess.domain.board.BoardFactory;
 import chess.domain.position.File;
 import chess.domain.position.Position;
 import chess.domain.position.Rank;
