+++ src/test/java/chess/domain/board/BoardFactoryTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import static chess.domain.piece.PieceType.BISHOP;
 import static chess.domain.piece.PieceType.EMPTY;
@@ -9,8 +9,8 @@ import static chess.domain.piece.PieceType.QUEEN;
 import static chess.domain.piece.PieceType.ROOK;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.Board;
-import chess.domain.BoardFactory;
+import chess.domain.board.Board;
+import chess.domain.board.BoardFactory;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
 import chess.domain.position.File;
